+++ lib/searchkick/reindex_v2_job.rb
@@ -6,6 +6,8 @@ module Searchkick
       model = klass.constantize
       # use should_index? to decide whether to index (not default scope)
       # just like saving inline
+      # could use Searchkick.scope() in future
+      # but keep for now for backwards compatibility
       model = model.unscoped if model.respond_to?(:unscoped)
       items = [{id: id, routing: routing}]
       model.searchkick_index.send(:record_indexer).reindex_items(model, items, method_name: method_name, single: true)