+++ test/models/artist.rb
@@ -1,0 +1,7 @@
+class Artist
+  searchkick unscope: true
+
+  def should_index?
+    should_index
+  end
+end